Move asset commands to dedicated `af assets` subcommand
- Create new `af assets` command group with list, events, triggers
- Remove asset commands from `af config`
- Cleaner CLI organization: assets are data-lineage related, not config
