+++ astro-airflow-mcp/src/astro_airflow_mcp/cli/config.py
@@ -221,115 +221,3 @@ def list_providers() -> None:
             output_json({"message": "No providers found", "response": data})
     except Exception as e:
         output_error(str(e))
-
-
-@app.command("assets")
-def list_assets(
-    limit: Annotated[
-        int,
-        typer.Option("--limit", "-l", help="Maximum number of assets to return"),
-    ] = 100,
-    offset: Annotated[
-        int,
-        typer.Option("--offset", "-o", help="Offset for pagination"),
-    ] = 0,
-) -> None:
-    """List data assets/datasets tracked by Airflow.
-
-    Returns asset information including URI, producing tasks,
-    and consuming DAGs for data lineage.
-    """
-    try:
-        adapter = get_adapter()
-        data = adapter.list_assets(limit=limit, offset=offset)
-
-        if "assets" in data:
-            result = wrap_list_response(data["assets"], "assets", data)
-            output_json(result)
-        else:
-            output_json({"message": "No assets found", "response": data})
-    except Exception as e:
-        output_error(str(e))
-
-
-@app.command("asset-events")
-def list_asset_events(
-    limit: Annotated[
-        int,
-        typer.Option("--limit", "-l", help="Maximum number of events to return"),
-    ] = 100,
-    offset: Annotated[
-        int,
-        typer.Option("--offset", "-o", help="Offset for pagination"),
-    ] = 0,
-    dag_id: Annotated[
-        str | None,
-        typer.Option("--dag-id", "-d", help="Filter by DAG that produced the event"),
-    ] = None,
-    run_id: Annotated[
-        str | None,
-        typer.Option("--run-id", "-r", help="Filter by DAG run that produced the event"),
-    ] = None,
-    task_id: Annotated[
-        str | None,
-        typer.Option("--task-id", "-t", help="Filter by task that produced the event"),
-    ] = None,
-) -> None:
-    """List asset/dataset events.
-
-    Asset events are produced when tasks update assets/datasets.
-    These events can trigger downstream DAGs (data-aware scheduling).
-
-    Examples:
-        af config asset-events
-        af config asset-events --dag-id my_dag
-        af config asset-events --dag-id my_dag --run-id run_123
-    """
-    try:
-        adapter = get_adapter()
-        data = adapter.list_asset_events(
-            limit=limit,
-            offset=offset,
-            source_dag_id=dag_id,
-            source_run_id=run_id,
-            source_task_id=task_id,
-        )
-
-        if "asset_events" in data:
-            result = wrap_list_response(data["asset_events"], "asset_events", data)
-            output_json(result)
-        else:
-            output_json({"message": "No asset events found", "response": data})
-    except Exception as e:
-        output_error(str(e))
-
-
-@app.command("asset-triggers")
-def get_upstream_asset_events(
-    dag_id: Annotated[str, typer.Argument(help="The DAG ID")],
-    dag_run_id: Annotated[str, typer.Argument(help="The DAG run ID")],
-) -> None:
-    """Get asset events that triggered a specific DAG run.
-
-    Shows which asset/dataset updates caused a DAG run to start.
-    Useful for understanding data-aware scheduling causation.
-
-    Examples:
-        af config asset-triggers my_dag scheduled__2024-01-01T00:00:00+00:00
-    """
-    try:
-        adapter = get_adapter()
-        data = adapter.get_dag_run_upstream_asset_events(dag_id, dag_run_id)
-
-        if "asset_events" in data:
-            result = {
-                "dag_id": dag_id,
-                "dag_run_id": dag_run_id,
-                "triggered_by_events": data["asset_events"],
-                "event_count": len(data["asset_events"]),
-            }
-            output_json(result)
-        else:
-            output_json(data)
-    except Exception as e:
-        output_error(str(e))
